Correction of Balancing
Illustrated the statement with the circuit of https://github.com/LibreSolar/bms-15s80-sc/blob/master/LibreSolar_BMS48V_schematic.pdf
I haven't seen a circuit bleeding current to common ground (0V reference), thus I question that even exists, as it would reduce the current charging the other cells? 

TODO: the image should be uploaded to the repo and should be .svg, adding some arrows [1] showing the current flow and maybe the chip switching the MOSFET would be a good addition. Nice to have would be  autogeneration of the illustration whenever the circuit changes. A link to the bms-15s80-sc could be added as well. 

[1] Resource: https://www.avdweb.nl/solar-bike/electronics/bms